Switch to 'int' valued popularities
It looks like my testing of fastjson output compatibility was incomplete
and the library generates float numbers different from what Jenkins core
writes out as the 'canonical' form (json-lib), resulting in checksum
errors.

Very small numbers are written in scientific notation in both json-lib
and fastjson, but fastjson uses 'E' while json-lib uses 'e'.

Floats without fractional part are written like ints by json-lib, while
fastjson writes them with '.0' suffix.

This basically means that, without switching out the serialization
library, we're unable to have floats in update center data.

+++ src/main/java/io/jenkins/update_center/PluginUpdateCenterEntry.java
@@ -177,7 +177,7 @@ public class PluginUpdateCenterEntry {
         return previous == null ? null : TIMESTAMP_FORMATTER.format(previous.getTimestamp());
     }
 
-    public float getPopularity() throws IOException {
+    public int getPopularity() throws IOException {
         return Popularities.getInstance().getPopularity(artifactId);
     }
